refactor: move NgView, NgElement and similar to separate module
