+++ nativescript-angular/directives/action-bar.ts
@@ -1,9 +1,11 @@
 import { Directive, Component, ElementRef, Optional, OnDestroy } from "@angular/core";
 import { ActionItem, ActionBar, NavigationButton } from "tns-core-modules/ui/action-bar";
-import { isBlank } from "../lang-facade";
 import { Page } from "tns-core-modules/ui/page";
 import { View } from "tns-core-modules/ui/core/view";
-import { registerElement, ViewClassMeta, NgView } from "../element-registry";
+
+import { isBlank } from "../lang-facade";
+import { registerElement } from "../element-registry";
+import { ViewClassMeta, NgView } from "../element-types";
 
 const actionBarMeta: ViewClassMeta = {
     skipAddToDom: true,
