+++ nativescript-angular/directives/tab-view.ts
@@ -9,7 +9,7 @@ import {
 } from "@angular/core";
 import { TabView, TabViewItem } from "tns-core-modules/ui/tab-view";
 
-import { DetachedElement } from "../element-registry";
+import { CommentNode } from "../element-types";
 import { convertToInt } from "../common/utils";
 import { rendererLog } from "../trace";
 import { isBlank } from "../lang-facade";
@@ -106,7 +106,7 @@ export class TabViewItemDirective implements OnInit {
         const viewRef = this.viewContainer.createEmbeddedView(this.templateRef);
         // Filter out text nodes and comments
         const realViews = viewRef.rootNodes.filter(node =>
-                            !(node instanceof DetachedElement));
+                            !(node instanceof CommentNode));
 
         if (realViews.length > 0) {
             this.item.view = realViews[0];
